Fix unwrapped null in getLatestMailForConversation()
The source of the problem is inconsistent state of selected items in
ListModel. ListModel has an issue with keeping track of selection with
ListAutoSelectBehavior.NEWER. When there isn't a newer item it will fall
back to selecting the first item from the list which might be the item
that we are removing, which means that there will be an item in the
selected list that is not in the overall items list (or more
specifically not in the filtered items list) which is a violation of
the contract. Because of that mismatch we would still try to populate
the selected items with conversation data even though we don't have it
anymore.

We fixed the issue by fixing the issue in ListModel itself. First we
remove the item from the list and only then we try to find candidate
for selection.

This also fixed the issue where the conversation is lingering in the
viewer even after being moved.

We also have changed getPreviousItem()/getNextItem() functions to take
the list so that it is explicit what they depend on.

Close #8717

Co-authored-by: wrd <wrd@tutao.de>