[Android] Cache Widget Events
This commits implements a simple in-memory cache to store fetched events
and just filter out events that have already finished instead of always
fetch from remote. With this, we can save network data and still have
the widget updated with only upcoming events.

Remote events are fetched once a day and when user login or
create/update/delete an event.