Fix toolbar actions for sticky mail
When opening a mail from a notification, some of the toolbar actions are
either noOp or not available. This is because we open the mail as a
sticky mail independently of the list state, resulting in selected mails
and subsequently actionable mails being an empty array.
Therefore, actions will either be hidden or do nothing.

Fixed by using conversation mails (primary mail when conversation view
is disabled) as actionable mails.

Close #8775

Co-authored-by: hrb-hub <hrb-hub@users.noreply.github.com>