[mimimi] fix invalid dates breaking the importer
The message parser returns -1 if it can't parse the Date: header.
Converting this with `n as u64` leads to an underflow and subsequently
to a panic because of the multiplication necessary to convert to
milliseconds.

We'll now use safe methods to default to 1970-01-01T00:00:00.000 if the
date is unusable.

fix #8766