Fix move selected mails from folder not working on mobile
Introduced in c65f0f22
Moving selected mails in a list was broken due to a small refactor in
`MailView`'s `moveMailsFromFolder`. The refactor changed when selected
mails are captured from before the `onSelected` action to after it.
Because the onSelected action is selectNone on mobile, we end up
capturing 0 mails for the move.

Reverting the refactor fixes this by capturing the selected mails before
showing the folder selection dropdown.

Fix #8891