fix entityUpdate handling when typeId is null in legacy case
Old entity updates do not have typeIds set on the EntityUpdateData.
However, In some places, this data is used to construct TypeRefs. We
have added a function to return the TypeRef corresponding to the type
using application and typeName on the EntityUpdate. The changes made in
this commit can be removed 45 days after server release and the schema
upgrade for the attrId refactoring as all entityUpdates will have
typeIds by then.

Co-authored-by: jomapp <17314077+jomapp@users.noreply.github.com>